+++ packages/datasource-toolkit/src/interfaces/action.ts
@@ -187,6 +187,12 @@ export type ActionFieldJsonEditor = ActionFieldBase & {
   widget: 'JsonEditor';
 };
 
+export type ActionFieldAddressAutocomplete = ActionFieldBase & {
+  type: 'String';
+  widget: 'AddressAutocomplete';
+  placeholder?: string;
+};
+
 export type ActionFieldDropdownAll =
   | ActionFieldDropdown<'Date' | 'Dateonly' | 'Number' | 'String' | 'StringList', string>
   | ActionFieldDropdown<'Number' | 'NumberList', number>;
@@ -221,6 +227,7 @@ export type ActionField = StrictUnion<
   | ActionFieldTimePicker
   | ActionFieldJsonEditor
   | ActionFieldFilePicker
+  | ActionFieldAddressAutocomplete
 >;
 
 export type ActionFieldWidget =
@@ -237,6 +244,7 @@ export type ActionFieldWidget =
   | 'NumberInputList'
   | 'ColorPicker'
   | 'DatePicker'
+  | 'AddressAutocomplete'
   | 'UserDropdown'
   | 'TimePicker'
   | 'FilePicker'