+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,30 @@
+## @forestadmin/agent [1.34.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.34.0...@forestadmin/agent@1.34.1) (2023-09-26)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.34.1
+
+# @forestadmin/agent [1.34.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.33.0...@forestadmin/agent@1.34.0) (2023-09-26)
+
+
+### Features
+
+* **widgets:** add support for address picker in custom actions ([#832](https://github.com/ForestAdmin/agent-nodejs/issues/832)) ([6441d8e](https://github.com/ForestAdmin/agent-nodejs/commit/6441d8ee257ef6bef3e27f18f37377eecf0c0730))
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.34.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.27.0
+* **@forestadmin/forestadmin-client:** upgraded to 1.23.0
+
 # @forestadmin/agent [1.33.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.32.0...@forestadmin/agent@1.33.0) (2023-09-25)